Lots of refactoring
DoubleField sigfigs now optional, set to 0 or lower to disable.
TextField and children now has KeyListener to get more frequent updates
of when the text changes.
Added hasInput() to TextField and children to determine if it has valid
content if any.
DoubleField and IntegerField don’t throw error if field is empty.